Add files via upload
Crazy story so yeas before I ever leaned how to code, I was sitting in Starbucks with a broken right hand and I came to  my brother with the idea of automating his costs tracking. Instinctually I wrote down a program for him but it was in pseudo code. My right hand was broken so I wrote it all down with my left hand. Thats how I became ambidextrous through hardship and necessity. Years pass and I find myself in computer science 101. Digging though my old stuff I found the skeleton sudo code template and I immediately knew how to code it. Here is my first iteration. We've tested it and won a few jobs of 40k and profited 8k on the job.